projetos/projeto-final-felipe_vitor-main/code/app/requirements.txt
altgraph==0.17.4
Nuitka==2.6.8
numpy==2.3.2
ordered-set==4.1.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.1
patchelf==0.17.2.4
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.7
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
tzdata==2025.2
zstandard==0.24.0